requirements.txt
flask==3.0.2
flask-login==0.6.3
google-generativeai==0.3.2
werkzeug==3.0.1
python-dotenv==1.0.0